chore: improve PseudoBox exports
